Safari-609.1.11.1/LayoutTests/accessibility/mac/attributed-string-includes-misspelled-with-selection-expected.txt

word mispelled word
This tests that if the cursor is on a misspelled word, the misspelled marker will still be returned for that range. The test places selection in the middle of 'mispelled' and then queries for the attributed string.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS attrString.indexOf('AXMisspelled = 1') != -1 is true
PASS successfullyParsed is true

TEST COMPLETE
